+++ shortcodes/legacy/file.py
@@ -9,6 +9,7 @@ class Shortcode():
 		self.description = "Processes the file content of 'path.'"
 
 	def run_atomic(self, pargs, kwargs, context):
+		self.log.warning(f"As of v9.14.0, [file] is a legacy shortcode and will eventually be removed in favor of [call] - the main difference is that [call] also works with functions.")
 		if "_bypass_if" in kwargs:
 			if self.Unprompted.parse_advanced(kwargs["_bypass_if"], context): return ""
 
@@ -43,10 +44,10 @@ class Shortcode():
 			if (key[0] == "_"): continue  # Skips system arguments
 			self.Unprompted.shortcode_objects["set"].run_block([key], {}, context, value)
 
-		self.Unprompted.shortcode_objects["else"].do_else = False
+		self.Unprompted.conditional_depth = 0
 		return (self.Unprompted.process_string(file_contents, path))
 
 	def ui(self, gr):
-		gr.Textbox(label="Filepath 🡢 str", max_lines=1)
+		gr.Textbox(label="Function name or filepath 🡢 str", max_lines=1)
 		gr.Textbox(label="Expected encoding 🡢 _encoding", max_lines=1, value="utf-8")
 		pass